[feat/#104] 타입 수정
