+++ src/components/FormDetails/FormDetailsInfo/FormDetailsInfo.tsx
@@ -1,4 +1,4 @@
-import { FormDetailsProps } from '@/lib/types/types'
+import { FormDetailsProps } from '@/lib/types/formTypes'
 import Image from 'next/image'
 import { useEffect, useState } from 'react'
