+++ src/lib/types/formTypes.ts
@@ -1,0 +1,38 @@
+export interface FormDetailsProps {
+  updatedAt?: string
+  createdAt?: string
+  preferred?: string
+  age?: string
+  education?: string
+  gender?: string
+  numberOfPositions?: number
+  isPublic?: boolean
+  hourlyWage?: number
+  workDays?: string[]
+  workEndTime?: string
+  workStartTime?: string
+  workEndDate?: string
+  workStartDate?: string
+  location?: string
+  imageUrls?: string[]
+  recruitmentEndDate?: string
+  recruitmentStartDate?: string
+  description?: string
+  title?: string
+  ownerId?: number
+  id?: number
+  scrapCount?: number
+  applyCount?: number
+  isScrapped?: boolean
+  phoneNumber?: string
+  storePhoneNumber?: string
+}
+
+export interface ApplicationProps {
+  id: number
+  applicantId: number
+  name: string
+  phoneNumber: string
+  experienceMonths: number
+  status: string
+}
