+++ src/lib/types/types.ts
@@ -67,33 +67,3 @@ export interface MessageContext {
 export interface ButtonText {
   signInOut: '로그인' | '로그아웃'
 }
-
-export interface FormDetailsProps {
-  updatedAt?: string
-  createdAt?: string
-  preferred?: string
-  age?: string
-  education?: string
-  gender?: string
-  numberOfPositions?: number
-  isPublic?: boolean
-  hourlyWage?: number
-  workDays?: string[]
-  workEndTime?: string
-  workStartTime?: string
-  workEndDate?: string
-  workStartDate?: string
-  location?: string
-  imageUrls?: string[]
-  recruitmentEndDate?: string
-  recruitmentStartDate?: string
-  description?: string
-  title?: string
-  ownerId?: number
-  id?: number
-  scrapCount?: number
-  applyCount?: number
-  isScrapped?: boolean
-  phoneNumber?: string
-  storePhoneNumber?: string
-}
